backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
PyJWT==2.8.0
structlog>=23.1.0

# Cache & Session Storage
redis>=5.0.1
aioredis>=2.0.1

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# API Clients
openai>=1.3.0
anthropic>=0.7.0
requests>=2.31.0
aiohttp>=3.9.0
aiofiles>=23.2.0

# Text Processing & NLP
sentence-transformers>=2.2.2
nltk>=3.8.1
textblob>=0.17.1
spacy>=3.7.2

# Vector Database
chromadb>=0.4.15

# User Agent & IP Analysis
user-agents>=2.2.0
ua-parser>=0.18.0
geoip2>=4.7.0
maxminddb>=2.5.1

# File Processing
python-magic>=0.4.27
pandas>=2.1.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# Report Generation
fpdf2>=2.7.6
reportlab>=4.0.7
jinja2>=3.1.2
weasyprint>=60.0

# Logging & Monitoring
python-json-logger>=2.0.7
structlog>=23.2.0

# HTTP & Web Utilities
httpx>=0.25.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.2

# Data Processing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Async & Concurrency
asyncio-throttle>=1.0.2
asyncpg>=0.29.0

# Validation & Parsing
validators>=0.22.0
python-dateutil>=2.8.2
pytz>=2023.3

# System & Performance Monitoring
psutil>=5.9.0

# Development & Testing (optional but useful)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.14.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Production Deployment
gunicorn>=21.2.0
supervisor>=4.2.5

# Rate Limiting
slowapi>=0.1.8

# CORS
fastapi-cors>=0.0.6

# Background Tasks
celery>=5.3.0
apscheduler>=3.10.4

# Metrics & Health Checks
prometheus-client>=0.19.0

# Email (for notifications)
emails>=0.6.0
sendgrid>=6.10.0

# Webhook handling
webhooks>=0.4.2

# CSV Processing
csv-reader>=1.2.0 
csvkit>=1.1.1

# Additional utilities that might be needed
click>=8.1.7  # For CLI commands
rich>=13.7.0  # For better console output
tqdm>=4.66.0  # Progress bars
schedule>=1.2.0  # Job scheduling

# MISSING PACKAGES IDENTIFIED FROM CODE ANALYSIS
# These were causing the ModuleNotFoundError crashes

# Standard library backports and enhancements
typing-extensions>=4.8.0
dataclasses>=0.8; python_version < "3.7"

# Concurrent processing (from test files)
concurrent-futures>=3.1.1; python_version < "3.2"

# Additional authentication
bcrypt>=4.0.1

# Machine Learning dependencies
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0

# Additional async utilities
anyio>=4.0.0

# Email validation
email-validator>=2.0.0

# Advanced cryptography
cryptography>=41.0.0

# JSON processing
ujson>=5.8.0

# Additional utilities
shortuuid>=1.0.11
arrow>=1.3.0


# Regular expressions
regex>=2023.10.3

# Collections utilities
collections-extended>=2.0.2

# Functional programming
toolz>=0.12.0

# URL processing
urllib3>=2.0.0

# Pathlib backport
pathlib2>=2.3.7; python_version < "3.4"

# Statistics backport
statistics>=1.0.3.5; python_version < "3.4"

# Additional testing utilities
mock>=5.1.0

# Caching utilities
cachetools>=5.3.2

# Retry utilities
tenacity>=8.2.3

# String utilities
unidecode>=1.3.7

# Text distance
textdistance>=4.6.0

# Fuzzy matching
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.1

# Language detection
langdetect>=1.0.9

# Image processing
Pillow>=10.0.0

# Encoding detection
chardet>=5.2.0

# File type detection
filetype>=1.2.0

# Compression
zstandard>=0.21.0

# Additional development tools
isort>=5.12.0
autoflake>=2.2.1

# Error tracking
sentry-sdk>=1.38.0

# Additional testing
factory-boy>=3.3.0
faker>=20.1.0
freezegun>=1.2.2

# WebSocket support
websockets>=11.0.3

# Additional database utilities
sqlmodel>=0.0.8

# Security token handling
itsdangerous>=2.1.2

# Additional monitoring
memory-profiler>=0.61.0

# Build tools (sometimes needed)
wheel>=0.41.0
setuptools>=68.0.0